refactor(platform): migrate v1-runs.ts logs handlers to mockapi helper
rename v1-runs.ts to api-logs.ts and migrate both GET /api/zero/logs and
GET /api/zero/logs/:id handlers from raw http.* MSW calls to the contract-driven
mockApi helper using the existing logsListContract and logsByIdContract from @vm0/core.

Closes #9946
